Release #1 (16.01.2017)

* was added simple google map into html page

* agent object, type, service, controller and mock simulator

* List of agents changed to Set, maps-init renamed to google-maps

* removed unused const

* fixed Agent comparator

* delete method for agents, update random generator, count of agents

* creating simple markers on map

* randomly appearing dots on map

* randomly appearing 20 agents on map

* showing way to destination on a map

* were added post and patch method for agent, removed redundant agent Service

* properties to yml, fixed typo in js

* moving marker on the map

* fixed speed coef

* hidden alert on google api error

* paginating response of all agents

* was added pagination depending on amount of agents,
    functionality for all agent moved to ManagementController

* changed simulation area

* Initial Bitbucket Pipelines configuration

* Initial Bitbucket Pipelines configuration

* controllers refactor, async simulation method

* smoothly moving markers

* separated pool of objects and simulator

* distance algorithm

* moving to random way

* changes in processing markers pagination on front

* basic osm lib

* finding static objects(bus stops, light and crossings)

* coloring ways on the map

* eliminated building roads

* controllers renamed to controller,
    maps module moved to web/resources/public,
    `rest` module renamed into `web`
    created MapController that should return `map.html`(currently did not),

* controllers renamed to controller

* removed unused include from settings.gradle

* extending `WebMvcConfigurerAdapted`, testing `HomeController` and `hello.html` remove it before merge

* returning html(not working in a proper way js and css aren't returned)

* added `@RequestMapping` for rest cotrollers, which fixes a bug

* html is returned by controller with all js and css files

* removed `extends WebMvcConfigurerAdapter` from main that is not necessary

* gradle config refactor,
    reading osm file from stream

* side menu, way filter

* customizing road that are showed

* optimized showing ways

* types of static objects on front

* display/hide left menu bar

* updated build version

* removed unused static var

* graph representation showed on the map

* showing markers on a map

* rescaling graph(road are shown with interceptions)

* minimizing graph

* updated application yml, graph correction, rescaling and minimaliztion

* connected vertices and avoid duplicate vertices

* refactor properties, graph connection

* rescaling graph(fixed version)

* added extra map files to parse

* renamed czrnowiejska into czarn

* cors config in java, auth for actuator

* CorsConfig Renamed into SecurityConfig, encrypted password for actuator

* updated build version

* updated build version

* rescale algorithm small refactor

* removed cors

* fixed actuator properties

* were added extra roadtypes, parsing locationRange,

* renamed map2 to osmmap controller

* map moved to czarnowiejska

* refactored controllers, simulator and Pool, fixing bugs in cachedTypes

* refactor in async simulation method, presenting simulation,

* weak simulation

* refactored way of finding next step

* fixed marker moving

* refactored connecting markers that are located closely

* changed car marker

* deleting agents that are at the edges of sim area

* initialized simulator that will work with lights

* changed algo of finding close vertexes

* killing agents that reached end of a map

* fixed bug with deleting dead vertices

* removed logger for empty rechable vertices and extra null checks

* updated build version

* js ajax replaced with jquery ajax

* refactor renamed const into common, initializing and printing list of static and way types in one place

* pagination moved to another file

* optimized agent dto

* was added error handler for ajax

* menu for showing graph options

* paginating requests

* implemented pagination

* refactored RoadType, added checker for type group

* checking if no start points

* added new, removed old osm files

* refactor, changed sim file

* added scroll for menu

* replacing REST_URL on release version

* changed println for gradle

* collecting and representing statistic

* fixed vertex agent incremention bug, add =ed uptime controller

* fixed rescaling bug

* fixed typo

* updated build version

* updated build version for milestone #2

* added service to ping running application in azure cloud

* updated build version

* updated build version

* logger config

* changed log file name

* reverted map file

* Merge branch 'develop' of https://bitbucket.org/ypiets/artificial-city-simulator into feature/logger

    # Conflicts:
    #	build.gradle
    #	web/src/main/resources/application.yml

* removed pagination

* improved logging config

* fixed log file property name

* fixed bug counting time from app start

* updated build version

* changed timeout updating agents

* * removed MockSimulator
    * renamed StaticSimulator to LightsSimulator
    * refactored AgentSimulator
    * LightsSimulator refactored to implement Simulator interface

* * created relation class
    * refactored LightsSimulator

* refactored simulations reset controller

* enabled async for simulations

* removed comments from SimulatorManagementController

* setting vertex for static point, added new todo's

* wtf: created LightsSimulator in a second the time

* basic lights relation

* fix: saving relation to light

* changed listener for removing statistic markers

* statistic controller returning only lights static points

* updating lights marker

* updated build version

* removed MockAgent, changed line formation

* Build version for first release.

* Merge branch 'develop' of https://bitbucket.org/ypiets/artificial-city-simulator into develop

    # Conflicts:
    #	build.gradle

* changed builder for agent, removed reangeDto, foramtion changes

* Build version for first release.

Approved-by: GM <nondorian@gmail.com>
Approved-by: Mateusz Raczyński <matrac7@gmail.com>

+++ simulation/src/main/java/com/acs/simulator/impl/AgentSimulatorImpl.java
@@ -6,13 +6,12 @@ import com.acs.models.graph.Vertex;
 import com.acs.models.statics.StaticPoint;
 import com.acs.pool.def.AgentPool;
 import com.acs.service.GraphService;
-import com.acs.simulator.def.Simulator;
+import com.acs.simulator.def.AgentSimulator;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Primary;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
 import java.util.Queue;
@@ -21,9 +20,8 @@ import java.util.Set;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
-@Primary
-@Component
-public class RSimulator implements Simulator {
+@Service
+public class AgentSimulatorImpl implements AgentSimulator {
 
     private final Logger logger = LogManager.getLogger(this.getClass());
 
@@ -34,13 +32,10 @@ public class RSimulator implements Simulator {
     private final Queue<Agent> queueToAppear = new LinkedBlockingQueue<>();
 
     @Autowired
-    public RSimulator(AgentPool pool,
-                      GraphService graphService,
-                      StaticSimulator staticSimulator) {
+    public AgentSimulatorImpl(AgentPool pool,
+                              GraphService graphService) {
         this.pool = pool;
         this.graphService = graphService;
-//        uncomment next line when bug with 0 vertices will be fixed
-//        staticSimulator.startLightsDaemon();
     }
 
     @PostConstruct
@@ -58,9 +53,9 @@ public class RSimulator implements Simulator {
 
     private void createRandomAgent(Set<Vertex> startVertices) {
         Vertex vertex = randomValueFromSet(startVertices);
-        if(vertex == null){
+        if (vertex == null) {
             vertex = randomValueFromSet(graphService.getGraph().getVertices());
-            if(vertex == null) {
+            if (vertex == null) {
                 return;
             }
         }
@@ -118,9 +113,9 @@ public class RSimulator implements Simulator {
             pool.kill(agent);
             return null;
         } else if (reachableVertices.size() == 1) {
-            Vertex vertex = reachableVertices.stream().findFirst().get();
+            Vertex vertex = reachableVertices.stream().findFirst().orElse(null);
 
-            if (vertex == agent.getVertex()) {
+            if (vertex != null && vertex == agent.getVertex()) {
                 vertex.setAgent(null);
                 pool.kill(agent);
                 return null;
@@ -210,7 +205,7 @@ public class RSimulator implements Simulator {
                     }
                 }
             } catch (InterruptedException e) {
-                e.printStackTrace();
+                logger.error("Unexpected: ", e);
             }
         }).start();
     }
